Create 1_arrow_key_press.tsx
I want to create a simple app with a number that can be changed on arrow key presses. My goal is to not change it on the key press but only change it after 1 second of holding the key. 
(+1 every second you hold the up arrow key, -1 every second you hold the down arrow key)
The problem with my code below is that the number starts changing really fast (not every sec as it should) and it doesnt stop even after i let go of the key.